+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Send metrics to Analytics (Create Quote and Send Message events)
-
+- Send use quote metrics to Analytics 
+  
 ### Removed
 - [ENGINEERS-1247] - Disable cypress tests in PR level